Add documentation for Authentication
A detailed documentation for Authentication has been added. It highlights the role of Trust over IP and NUTS in achieving authenticity in information exchange within healthcare. It further elaborates on sources of trust such as UZI and URA credentials, as well as a possible practical implementation considering the current lack of official trusted sources.