Refactor build_deploy workflow
This commit simplifies the GitHub Actions build_deploy.yml file. It replaces the previous complex process of logging in to the Container registry, building, and pushing an image. Instead, it implements a simplified build and push to a local Docker registry. In addition, it also adds a step to inspect the results with a distinctly named Docker container.