Update Docker run command in GitHub workflow
The Docker run command in the GitHub Actions workflow is updated to use a dynamic current directory path for volume binding. Previously, static paths were used which could lead to errors if the workflow is run from a different location. Now, it automatically picks up from the current working directory, increasing the portability and reliability of the script.